feat: implement new Action buttons
