+++ src/components/Action/index.tsx
@@ -1,1 +1,2 @@
-export { USAction as Action } from './Action';
+export { Action } from './Action';
+export { USAction as OldAction } from './OldAction';
