+++ src/components/Modal/Modal.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, useWindowDimensions, View } from 'react-native';
 import { Body } from '../Body';
-import { Action } from '../Action';
+import { OldAction } from '../Action';
 import type { TourStep } from '../../types';
 import { useMemo } from 'react';
 import { useStore } from '../../stores/useStore';
@@ -59,7 +59,7 @@ export const Modal = ({ step }: ModalProps) => {
       </View>
       <View style={styles.modalActions}>
         {actions.map((action) => {
-          return <Action key={action.id} {...{ action }} />;
+          return <OldAction key={action.id} {...{ action }} />;
         })}
       </View>
       <RenderProgressBar />
